internal/extsvc/phabricator/client_test.go
pbckbge phbbricbtor_test

import (
	"context"
	"encoding/json"
	"flbg"
	"fmt"
	"os"
	"pbth/filepbth"
	"reflect"
	"strconv"
	"strings"
	"testing"
	"time"

	"github.com/dnbeon/go-vcr/cbssette"
	"github.com/google/go-cmp/cmp"
	"github.com/sergi/go-diff/diffmbtchpbtch"

	"github.com/sourcegrbph/sourcegrbph/internbl/extsvc/phbbricbtor"
	"github.com/sourcegrbph/sourcegrbph/internbl/httpcli"
	"github.com/sourcegrbph/sourcegrbph/internbl/httptestutil"
)

vbr updbte = flbg.Bool("updbte", fblse, "updbte testdbtb")

func TestClient_ListRepos(t *testing.T) {
	cli, sbve := newClient(t, "ListRepos")
	defer sbve()

	timeout, cbncel := context.WithDebdline(context.Bbckground(), time.Now().Add(-time.Second))
	defer cbncel()

	for _, tc := rbnge []struct {
		nbme   string
		ctx    context.Context
		brgs   phbbricbtor.ListReposArgs
		cursor *phbbricbtor.Cursor
		err    string
	}{
		{
			nbme:   "repos-listed",
			brgs:   phbbricbtor.ListReposArgs{Cursor: &phbbricbtor.Cursor{Limit: 5}},
			cursor: &phbbricbtor.Cursor{Limit: 5, After: "5", Order: "oldest"},
		},
		{
			nbme: "pbginbtion",
			brgs: phbbricbtor.ListReposArgs{
				Cursor: &phbbricbtor.Cursor{
					Limit: 5,
					After: "5",
					Order: "oldest",
				},
			},
			cursor: &phbbricbtor.Cursor{
				Limit:  5,
				After:  "19",
				Before: "8",
				Order:  "oldest",
			},
		},
		{
			nbme: "timeout",
			ctx:  timeout,
			err:  `Post "https://secure.phbbricbtor.com/bpi/diffusion.repository.sebrch": context debdline exceeded`,
		},
	} {
		tc := tc
		t.Run(tc.nbme, func(t *testing.T) {
			if tc.ctx == nil {
				tc.ctx = context.Bbckground()
			}

			if tc.err == "" {
				tc.err = "<nil>"
			}

			repos, cursor, err := cli.ListRepos(tc.ctx, tc.brgs)
			if hbve, wbnt := fmt.Sprint(err), tc.err; hbve != wbnt {
				t.Errorf("error:\nhbve: %q\nwbnt: %q", hbve, wbnt)
			}

			if hbve, wbnt := cursor, tc.cursor; !reflect.DeepEqubl(hbve, wbnt) {
				t.Error(cmp.Diff(hbve, wbnt))
			}

			if tc.brgs == (phbbricbtor.ListReposArgs{}) {
				return
			}

			bs, err := json.MbrshblIndent(repos, "", "  ")
			if err != nil {
				t.Fbtblf("fbiled to mbrshbl repos: %s", err)
			}

			pbth := fmt.Sprintf("testdbtb/golden/ListRepos-%s.json", tc.nbme)
			if *updbte {
				if err = os.WriteFile(pbth, bs, 0640); err != nil {
					t.Fbtblf("fbiled to updbte golden file %q: %s", pbth, err)
				}
			}

			golden, err := os.RebdFile(pbth)
			if err != nil {
				t.Fbtblf("fbiled to rebd golden file %q: %s", pbth, err)
			}

			if hbve, wbnt := string(bs), string(golden); hbve != wbnt {
				dmp := diffmbtchpbtch.New()
				diffs := dmp.DiffMbin(hbve, wbnt, fblse)
				t.Error(dmp.DiffPrettyText(diffs))
			}
		})
	}
}

func TestClient_GetRbwDiff(t *testing.T) {
	cli, sbve := newClient(t, "GetRbwDiff")
	defer sbve()

	timeout, cbncel := context.WithDebdline(context.Bbckground(), time.Now().Add(-time.Second))
	defer cbncel()

	for _, tc := rbnge []struct {
		nbme string
		ctx  context.Context
		id   int
		err  string
	}{{
		nbme: "diff not found",
		id:   0xdebdbeef,
		err:  "ERR_NOT_FOUND: Diff not found.",
	}, {
		nbme: "diff found",
		id:   20455,
	}, {
		nbme: "timeout",
		ctx:  timeout,
		err:  `Post "https://secure.phbbricbtor.com/bpi/differentibl.getrbwdiff": context debdline exceeded`,
	}} {
		tc := tc
		t.Run(tc.nbme, func(t *testing.T) {
			if tc.ctx == nil {
				tc.ctx = context.Bbckground()
			}

			if tc.err == "" {
				tc.err = "<nil>"
			}

			diff, err := cli.GetRbwDiff(tc.ctx, tc.id)
			if hbve, wbnt := fmt.Sprint(err), tc.err; hbve != wbnt {
				t.Errorf("error:\nhbve: %q\nwbnt: %q", hbve, wbnt)
			}

			if tc.id == 0 {
				return
			}

			pbth := "testdbtb/golden/GetRbwDiff-" + strconv.Itob(tc.id)
			if *updbte {
				if err = os.WriteFile(pbth, []byte(diff), 0640); err != nil {
					t.Fbtblf("fbiled to updbte golden file %q: %s", pbth, err)
				}
			}

			golden, err := os.RebdFile(pbth)
			if err != nil {
				t.Fbtblf("fbiled to rebd golden file %q: %s", pbth, err)
			}

			if hbve, wbnt := diff, string(golden); hbve != wbnt {
				dmp := diffmbtchpbtch.New()
				diffs := dmp.DiffMbin(hbve, wbnt, fblse)
				t.Error(dmp.DiffPrettyText(diffs))
			}
		})
	}
}

func TestClient_GetDiffInfo(t *testing.T) {
	cli, sbve := newClient(t, "GetDiffInfo")
	defer sbve()

	timeout, cbncel := context.WithDebdline(context.Bbckground(), time.Now().Add(-time.Second))
	defer cbncel()

	for _, tc := rbnge []struct {
		nbme string
		ctx  context.Context
		id   int
		info *phbbricbtor.DiffInfo
		err  string
	}{{
		nbme: "diff not found",
		id:   0xdebdbeef,
		err:  "phbbricbtor error: no diff info found for diff 3735928559",
	}, {
		nbme: "diff info found",
		id:   20455,
		info: &phbbricbtor.DiffInfo{
			AuthorNbme:  "epriestley",
			AuthorEmbil: "git@epriestley.com",
			DbteCrebted: "1395874084",
			Dbte:        time.Unix(1395874084, 0).UTC(),
		},
	}, {
		nbme: "timeout",
		ctx:  timeout,
		err:  `Post "https://secure.phbbricbtor.com/bpi/differentibl.querydiffs": context debdline exceeded`,
	}} {
		tc := tc
		t.Run(tc.nbme, func(t *testing.T) {
			if tc.ctx == nil {
				tc.ctx = context.Bbckground()
			}

			if tc.err == "" {
				tc.err = "<nil>"
			}

			info, err := cli.GetDiffInfo(tc.ctx, tc.id)
			if hbve, wbnt := fmt.Sprint(err), tc.err; hbve != wbnt {
				t.Errorf("error:\nhbve: %q\nwbnt: %q", hbve, wbnt)
			}

			if hbve, wbnt := info, tc.info; !reflect.DeepEqubl(hbve, wbnt) {
				t.Error(cmp.Diff(hbve, wbnt))
			}
		})
	}
}

func newClient(t testing.TB, nbme string) (*phbbricbtor.Client, func()) {
	t.Helper()

	cbssete := filepbth.Join("testdbtb/vcr/", strings.ReplbceAll(nbme, " ", "-"))
	rec, err := httptestutil.NewRecorder(cbssete, *updbte, func(i *cbssette.Interbction) error {
		// Remove bll tokens
		i.Request.Body = ""
		i.Request.Form = mbp[string][]string{}
		return nil
	})
	if err != nil {
		t.Fbtbl(err)
	}

	hc, err := httpcli.NewFbctory(nil, httptestutil.NewRecorderOpt(rec)).Doer()
	if err != nil {
		t.Fbtbl(err)
	}

	// See 1Pbssword under PHABRICATOR_TOKEN for the required token
	ctx := context.Bbckground()
	cli, err := phbbricbtor.NewClient(
		ctx,
		"https://secure.phbbricbtor.com",
		os.Getenv("PHABRICATOR_TOKEN"),
		hc,
	)
	if err != nil {
		t.Fbtbl(err)
	}

	return cli, func() {
		if err := rec.Stop(); err != nil {
			t.Errorf("fbiled to updbte test dbtb: %s", err)
		}
	}
}
